Follow checkstyle conventions
