+++ src/main/java/seedu/address/model/Model.java
@@ -124,7 +124,8 @@ public interface Model {
     /**
      * Replaces the given patient {@code target} with {@code editedPatient}.
      * {@code target} must exist in the address book.
-     * The patient identity of {@code editedPatient} must not be the same as another existing patient in the address book.
+     * The patient identity of {@code editedPatient} must not be the same
+     * as another existing patient in the address book.
      */
     void setPatient(Patient target, Patient editedPatient);
